Services/Tempo/gcp_deployment/requirements.txt
numpy
xarray
matplotlib
cartopy
harmony-py
pandas
redis
psycopg2-binary
pg8000
flask
gunicorn
netcdf4